Event Manager Created, Ship properties added,basic game setup. and more
Some basic setup (mainly for testing) for players and grid size are
implemented.
The Event Manager is created, which will handle turns and the main
loop/order of events in the game (we want to keep that out of main when
possible).  Ships now have health and damage.  Some values (which may
need changing) are set in the constructor for each ship type.  Players
now have their own grid, which will be that player's perspective/view.
Players have a ship of each type, a satellite, and a vector of fighters
and bombers.  A couple new properties (health and damage) were added to
the Aircraft Class.